requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain-mistralai